requirements.txt
# Activities API - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# HTTP Client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.12.0
ruff==0.1.8
mypy==1.7.1
